# backend/openapi.yaml
components:
  schemas:
...
    controller.apiResponse-any:
      properties:
        data: {}
        error:
          type: string
      type: object
...
    controller.apiResponse-array_model_Score:
      properties:
        data:
          items:
            $ref: '#/components/schemas/model.Score'
          type: array
          uniqueItems: false
        error:
          type: string
      type: object
    controller.apiResponse-array_model_ScoreAggregate:
      properties:
        data:
          items:
            $ref: '#/components/schemas/model.ScoreAggregate'
          type: array
          uniqueItems: false
        error:
          type: string
      type: object
...
    controller.apiResponse-model_Score:
      properties:
        data:
          $ref: '#/components/schemas/model.Score'
        error:
          type: string
      type: object
    controller.apiResponse-model_SystemEnrichment:
      properties:
        data:
          $ref: '#/components/schemas/model.SystemEnrichment'
        error:
          type: string
      type: object
...
    model.AuditRef:
      properties:
        email:
          type: string
        name:
          type: string
        role:
          type: string
        userid:
          type: string
      type: object
...
    model.FunctionOption:
      properties:
        description:
          type: string
        functionid:
          type: integer
        functionoptionid:
          type: integer
        optionname:
          type: string
        score:
          type: integer
      type: object
...
    model.PillarScore:
      properties:
        pillar:
          type: string
        pillarid:
          type: integer
        score:
          type: number
        tier:
          type: string
      type: object
...
    model.Score:
      properties:
        datacallid:
          type: integer
        datecalculated:
          type: number
        fismasystemid:
          type: integer
        functionoption:
          $ref: '#/components/schemas/model.FunctionOption'
        functionoptionid:
          type: integer
        last_edited_at:
          type: string
        last_edited_by:
          $ref: '#/components/schemas/model.AuditRef'
        notes:
          type: string
        scoreid:
          type: integer
      type: object
    model.ScoreAggregate:
      properties:
        datacallid:
          type: integer
        fismasystemid:
          type: integer
        pillarscores:
          items:
            $ref: '#/components/schemas/model.PillarScore'
          type: array
          uniqueItems: false
        systemscore:
          type: number
        systemtier:
          type: string
      type: object
    model.SystemEnrichment:
      properties:
        fisma_uuid:
          type: string
        payload:
          type: object
        synced_at:
          type: string
      type: object
...
  securitySchemes:
    bearerAuth:
      in: header
      name: Authorization
      type: apiKey
...
info:
  description: API for the Zero Trust Maturity Framework (ZTMF).
  title: ZTMF API
  version: 1.0.0
openapi: 3.1.0
paths:
...
  /scores:
    get:
      parameters:
      - description: Filter by FISMA system ID
        in: query
        name: fismasystemid
        schema:
          type: integer
      - description: Filter by data call ID
        in: query
        name: datacallid
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-array_model_Score'
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Internal Server Error
      security:
      - bearerAuth: []
      summary: List all scores
      tags:
      - scores
    post:
      parameters:
      - description: Score ID
        in: path
        name: scoreid
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/model.Score'
                description: Score to save
                summary: body
        description: Score to save
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-model_Score'
          description: Created
        "204":
          description: No Content
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Bad Request
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Internal Server Error
      security:
      - bearerAuth: []
      summary: Create or update a score
      tags:
      - scores
  /scores/{scoreid}:
    put:
      parameters:
      - description: Score ID
        in: path
        name: scoreid
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/model.Score'
                description: Score to save
                summary: body
        description: Score to save
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-model_Score'
          description: Created
        "204":
          description: No Content
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Bad Request
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Internal Server Error
      security:
      - bearerAuth: []
      summary: Create or update a score
      tags:
      - scores
  /scores/aggregate:
    get:
      parameters:
      - description: Filter by FISMA system ID
        in: query
        name: fismasystemid
        schema:
          type: integer
      - description: Filter by data call ID
        in: query
        name: datacallid
        schema:
          type: integer
      - description: Include per-pillar scores
        in: query
        name: include_pillars
        schema:
          type: boolean
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-array_model_ScoreAggregate'
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Internal Server Error
      security:
      - bearerAuth: []
      summary: Get aggregated scores
      tags:
      - scores
  /systemenrichment/{fisma_uuid}:
    get:
      parameters:
      - description: FISMA system UUID
        in: path
        name: fisma_uuid
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-model_SystemEnrichment'
          description: OK
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/controller.apiResponse-any'
          description: Internal Server Error
      security:
      - bearerAuth: []
      summary: Get enrichment payload for a FISMA system
      tags:
      - systemenrichment
...
servers:
- url: /api/v1